+++ native/src/include/cs_restbl.h
@@ -9,7 +9,7 @@
 
 using namespace oepd;
 
-EXP restbl::RESTBL* FromBinary(u8* src, int src_len);
+EXP restbl::RESTBL* FromBinary(u8* src, size_t src_len);
 EXP std::vector<u8>* ToBinary(restbl::RESTBL* handle);
 
 EXP bool Free(restbl::RESTBL* handle);
